fix: 문의하기 디자인 및 작동오류 수정 (#426)
