+++ pubspec.yaml
@@ -68,13 +68,14 @@ dependencies:
   shimmer: ^3.0.0
   firebase_auth: ^5.4.1
   sign_in_with_apple: ^6.1.4
-  mailer: ^6.0.1
+  mailer: ^6.6.0
   flutter_drawing_board: ^0.5.0
   path_provider: ^2.0.5
   photo_view: ^0.14.0
   win32: ^5.10.0
   package_info_plus: ^9.0.0
   url_launcher_ios: ^6.3.3
+  flutter_dotenv: ^6.0.0
 
 dev_dependencies:
   flutter_test:
@@ -99,6 +100,7 @@ flutter:
 
   # To add assets to your application, add an assets section, like this:
   assets:
+    - .env
     - assets/
     - assets/images/
     - assets/icons/
